Change to slightly awkward wording
In the existing text, the use of the phrase 'view more details about it' doesn't make sense in context, as what (if any) details would be visible before choosing an application are not stated. This change makes this sentence more clear in context and saves a few words in the process.